refacto: move yaml import to component
